backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-multitenant==3.0.0
psycopg2-binary==2.9.9
drf-yasg==1.21.7
whitenoise==6.6.0
django-jazzmin==3.0.4
